使用三种方式实现多线程带来的问题
