+++ src/main/java/com/github/hcsp/multithread/Counter2.java
@@ -1,0 +1,26 @@
+package com.github.hcsp.multithread;
+
+/**
+ * Third method to solve this problem
+ * Use synchronized
+ */
+
+public class Counter2 {
+    private int value = 0;
+
+    public int getValue() {
+        return value;
+    }
+
+    // 加上一个整数i，并返回加之后的结果
+    public synchronized int addAndGet(int i) {
+        value += i;
+        return value;
+    }
+
+    // 减去一个整数i，并返回减之后的结果
+    public synchronized int minusAndGet(int i) {
+        value -= i;
+        return value;
+    }
+}
